inline setup
