+++ apps/website/src/routes/docs/headless/combobox/examples/inline.tsx
@@ -1,0 +1,36 @@
+import { component$, useStyles$ } from '@builder.io/qwik';
+import { Combobox } from '@qwik-ui/headless';
+
+export default component$(() => {
+  useStyles$(styles);
+
+  const fruits = [
+    'Apple',
+    'Apricot',
+    'Bilberry',
+    'Blackberry',
+    'Blackcurrant',
+    'Currant',
+    'Cherry',
+    'Coconut',
+  ];
+
+  return (
+    <Combobox.Root class="combobox-root">
+      <Combobox.Label class="combobox-label">Personal Trainers</Combobox.Label>
+      <Combobox.Control class="combobox-control">
+        <Combobox.Input class="combobox-input" />
+      </Combobox.Control>
+      <Combobox.Inline>
+        {fruits.map((fruit) => (
+          <Combobox.Item key={fruit} class="combobox-item">
+            <Combobox.ItemLabel>{fruit}</Combobox.ItemLabel>
+          </Combobox.Item>
+        ))}
+      </Combobox.Inline>
+    </Combobox.Root>
+  );
+});
+
+// internal
+import styles from '../snippets/combobox.css?inline';
