+++ apps/website/src/routes/docs/headless/combobox/index.mdx
@@ -322,6 +322,10 @@ When the `<Combobox.ErrorMessage>` component is rendered, the component is in an
 
 > This allows you to use Qwik UI with any form validation library.
 
+### Inline Combobox
+
+<Showcase name="inline" />
+
 ## Example CSS
 
 Every code example uses the following CSS:
